Add `DayBarsCollector`
