+++ tradepy/collectors/adjust_factor.py
@@ -4,7 +4,7 @@ from tqdm import tqdm
 
 import tradepy
 from tradepy import LOG
-from tradepy.collectors import DataCollector
+from tradepy.collectors.base import DataCollector
 from tradepy.depot.misc import AdjustFactorDepot
