+++ tradepy/collectors/market_index.py
@@ -1,7 +1,7 @@
 import tradepy
 import pandas as pd
 from tradepy import LOG
-from tradepy.collectors import DataCollector
+from tradepy.collectors.base import DataCollector
 from tradepy.depot.index import SectorIndexBarsDepot, BroadBasedIndexBarsDepot
 from tradepy.conversion import broad_index_code_name_mapping
