+++ tradepy/collectors/stock_listing.py
@@ -2,7 +2,7 @@ import pandas as pd
 
 import tradepy
 from tradepy import LOG
-from tradepy.collectors import DataCollector
+from tradepy.collectors.base import DataCollector
 from tradepy.depot.stocks import StockListingDepot
