+++ tradepy/collectors/stock_minute_bars.py
@@ -2,7 +2,7 @@ import datetime
 
 import pandas as pd
 import tradepy
-from tradepy.collectors import DataCollector
+from tradepy.collectors.base import DataCollector
 from tradepy.utils import get_latest_trade_date
 from tradepy.depot.stocks import StockMinuteBarsDepot
