+++ tradepy/conversion.py
@@ -236,13 +236,13 @@ def convert_etf_current_quote(df: pd.DataFrame) -> pd.DataFrame:
         "最低价": "low",
         "开盘价": "open",
         "换手率": "turnover",
-        "总市值": "mkcap",
+        "总市值": "mkt_cap",
     }
     df.rename(columns=mapping, inplace=True)
     df = df[list(mapping.values())]
-    df["mkcap"] *= 1e-8  # convert to 100 mils
-    df["mkcap"] = df["mkcap"].round(3)
-    df.sort_values("mkcap", inplace=True)
+    df["mkt_cap"] *= 1e-8  # convert to 100 mils
+    df["mkt_cap"] = df["mkt_cap"].round(3)
+    df.sort_values("mkt_cap", inplace=True)
     return df
