+++ tradepy/depot/__init__.py
@@ -1,0 +1,4 @@
+from .etf import ETFListingDepot, ETFDailyBarsDepot
+from .misc import AdjustFactorDepot, RestrictedSharesReleaseDepot
+from .stocks import StockListingDepot, StocksDailyBarsDepot, StockMinuteBarsDepot
+from .index import BroadBasedIndexBarsDepot, SectorIndexBarsDepot
