+++ tradepy/depot/stocks.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 import tradepy
 from tradepy.conversion import convert_code_to_market
-from tradepy.depot.bases import GenericBarsDepot, GenericListingDepot
+from tradepy.depot.base import GenericBarsDepot, GenericListingDepot
 from tradepy.depot.misc import AdjustFactorDepot
 from tradepy.utils import get_latest_trade_date
